requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
xlsxwriter==3.1.9
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pdf2docx==0.5.6 